requirements.txt
selenium
webdriver-manager
<<<<<<< HEAD
gitpython
numpy 
matplotlib
pandas
bs4
pymongo
datetime
dnspython
ipython
pip
=======
gitPython
numpy==1.23.1
pandas==1.4.3
pymongo
datetime
beautifulsoup4
matplotlib
seaborn
forex-python
>>>>>>> 7a6a857929766f56301196042efbc499144b85ff
